トップページ以外のヘッダーにも各ページのリンク追加。
ついでにランキングページのレイアウト修正。
