+++ client/src/components/organisms/Header.tsx
@@ -1,37 +1,10 @@
 import { memo } from "react";
 import { Box, Flex, Image, Text } from "@chakra-ui/react";
-import { Link } from "react-router-dom";
 
 import { PrimaryContainer } from "../atoms/PrimaryContainer";
 
-import ranking from "../../image/rank1.png";
-import guild from "../../image/guild.png";
-import report from "../../image/report.png";
 import { CheerButton } from "../molcules/CheerButton";
-
-type Item = {
-  itemName: string;
-  src: string;
-  path: string;
-};
-
-const headerItems: Array<Item> = [
-  {
-    itemName: "ランキング",
-    src: ranking,
-    path: "/top/ranking",
-  },
-  {
-    itemName: "ギルド",
-    src: guild,
-    path: "/top/guild",
-  },
-  {
-    itemName: "レポート",
-    src: report,
-    path: "/top/report",
-  },
-];
+import { HeaderItems } from "../molcules/HeaderItems";
 
 type Props = {
   onOpenProfile: () => void;
@@ -68,21 +41,7 @@ export const Header = memo((props: Props) => {
         </Text>
         <Flex align="center" justify="space-between">
           <CheerButton onOpenCheer={onOpenCheer} />
-          {headerItems.map((item) => (
-            <Flex
-              key={item.itemName}
-              align="center"
-              justify="center"
-              flexDir="column"
-              ml={{ base: 6, md: 10 }}
-              color="whiteAlpha.800"
-              cursor="pointer"
-            >
-              <Link to={item.path}>
-                <Image src={item.src} w={{ base: "30px", md: "40px" }} />
-              </Link>
-            </Flex>
-          ))}
+          <HeaderItems />
         </Flex>
       </Flex>
     </PrimaryContainer>
