+++ client/src/components/organisms/HeaderGoBack.tsx
@@ -1,9 +1,11 @@
-import { Box } from "@chakra-ui/react";
 import { memo, useCallback } from "react";
+import { Box, Flex } from "@chakra-ui/react";
+import { Link, useHistory } from "react-router-dom";
 
 import { PrimaryContainer } from "../atoms/PrimaryContainer";
 import { SecondaryButton } from "../atoms/forms/SecondaryButton";
-import { useHistory } from "react-router-dom";
+import { HeaderItems } from "../molcules/HeaderItems";
+import HomeIcon from "@material-ui/icons/Home";
 
 export const HeaderGoBack = memo(() => {
   const history = useHistory();
@@ -14,8 +16,20 @@ export const HeaderGoBack = memo(() => {
 
   return (
     <PrimaryContainer>
-      <Box color="white" py={{ base: 3, md: 5 }} px={{ base: 5, md: 8 }}>
-        <SecondaryButton onClick={onClickGoBack}>戻る</SecondaryButton>
+      <Box py={{ base: 3, md: 5 }} px={{ base: 5, md: 8 }}>
+        <Flex align="center" justify="space-between">
+          <Box color="white">
+            <SecondaryButton onClick={onClickGoBack}>戻る</SecondaryButton>
+          </Box>
+          <Flex justify="space-between" align="center">
+            <Box fontSize={{ base: "27px", md: "36px" }} color="#d0cab7">
+              <Link to="/top">
+                <HomeIcon fontSize="inherit" color="inherit" />
+              </Link>
+            </Box>
+            <HeaderItems />
+          </Flex>
+        </Flex>
       </Box>
     </PrimaryContainer>
   );
